Move some blockchain utils to sandbox.py
